requirements.txt
peewee==3.14.0
discord.py~=1.5
